+++ node/clients/index.ts
@@ -6,6 +6,10 @@ import { Checkout } from './checkout'
 import { PvtCheckout } from './PvtCheckout'
 import { MasterData } from './masterdata'
 import { ProfileClient } from './profile'
+import { ProfileClientV1 } from './profile/profileV1'
+import { ProfileClientV2US } from './profile/profileV2US'
+import { ProfileClientV2EU } from './profile/profileV2EU'
+import { LicenseManagerExtendedClient } from './licenseManagerExtended'
 import { OMS } from './oms'
 import { Portal } from './portal'
 import { LogisticsClient } from './logistics'
@@ -35,6 +39,22 @@ export class Clients extends IOClients {
 
   public get profile() {
     return this.getOrSet('profile', ProfileClient)
+  }
+
+  public get profileV1() {
+    return this.getOrSet('profileV1', ProfileClientV1)
+  }
+
+  public get profileV2EU() {
+    return this.getOrSet('profileV2EU', ProfileClientV2EU)
+  }
+
+  public get profileV2US() {
+    return this.getOrSet('profileV2US', ProfileClientV2US)
+  }
+
+  public get licenseManagerExtended() {
+    return this.getOrSet('licenseManagerExtended', LicenseManagerExtendedClient)
   }
 
   public get oms() {