+++ node/resolvers/profile/services.ts
@@ -17,7 +17,7 @@ export async function getProfile(context: Context, customFields?: string) {
     : `profilePicture,id`
 
   return profile
-    .getProfileInfo(currentProfile, extraFields)
+    .getProfileInfo(currentProfile, context, extraFields)
     .then((profileData) => {
       if (profileData) {
         return {
@@ -51,13 +51,13 @@ export function getPasswordLastUpdate(context: Context) {
   }).then((response: any) => response.data.passwordLastUpdate)
 }
 
-export function getAddresses(context: Context) {
+export function getAddresses(context: Context, currentUserProfile?: Profile) {
   const {
     clients: { profile },
     vtex: { currentProfile },
   } = context
 
-  return profile.getUserAddresses(currentProfile)
+  return profile.getUserAddresses(currentProfile, context, currentUserProfile)
 }
 
 export async function getPayments(context: Context) {
@@ -66,7 +66,7 @@ export async function getPayments(context: Context) {
     vtex: { currentProfile },
   } = context
 
-  const paymentsRawData = await profile.getUserPayments(currentProfile)
+  const paymentsRawData = await profile.getUserPayments(currentProfile, context)
 
   if (!paymentsRawData?.paymentData) {
     return null
@@ -78,7 +78,7 @@ export async function getPayments(context: Context) {
   return availableAccounts.map((account: any) => {
     const { bin, availableAddresses, accountId, ...cleanAccount } = account
     const accountAddress = addresses.find(
-      (addr) => addr.addressName === availableAddresses[0]
+      (addr: Address) => addr.addressName === availableAddresses[0]
     )
 
     return { ...cleanAccount, id: accountId, address: accountAddress }
@@ -111,6 +111,7 @@ export async function updateProfile(
     .updateProfileInfo(
       currentProfile,
       newData,
+      context,
       extraFields && extraFields.customFieldsStr
     )
     .then(() => getProfile(context, extraFields && extraFields.customFieldsStr))
@@ -135,7 +136,7 @@ export function createAddress(context: Context, address: AddressInput) {
   const addressesData = mapNewAddressToProfile(address, currentProfile)
 
   return profile
-    .updateAddress(currentProfile, addressesData)
+    .updateAddress(currentProfile, addressesData, context)
     .then(() => getProfile(context))
 }
 
@@ -146,7 +147,7 @@ export function deleteAddress(context: Context, addressName: string) {
   } = context
 
   return profile
-    .deleteAddress(currentProfile, addressName)
+    .deleteAddress(currentProfile, addressName, context)
     .then(() => getProfile(context))
 }
 
@@ -168,7 +169,7 @@ export function updateAddress(
   })
 
   return profile
-    .updateAddress(currentProfile, addressesData)
+    .updateAddress(currentProfile, addressesData, context)
     .then(() => getProfile(context))
 }
 
@@ -195,12 +196,23 @@ export async function saveAddress(
   const addressesData = mapNewAddressToProfile(args.address, currentProfile)
   const [newId] = Object.keys(addressesData)
 
-  await profile.updateAddress(currentProfile, addressesData)
+  const result = await profile.updateAddress(
+    currentProfile,
+    addressesData,
+    context
+  )
 
-  const currentAddresses = await profile.getUserAddresses(currentProfile)
+  if (result && result.document) {
+    return result.document
+  }
+
+  const currentAddresses = await profile.getUserAddresses(
+    currentProfile,
+    context
+  )
 
   return currentAddresses.find(
-    (address) => address.addressName === newId
+    (address: Address) => address.addressName === newId
   ) as Address
 }